==> s05.Microbiome_population_structure.Rmd <==
---
title: "Microbiome population structure"
author: "Daoming Wang"
date: "2021/9/30"
output:
  html_document: 
    theme: flatly
    highlight: espresso
    toc: true
    toc_depth: 4
    toc_float: true
  word_document: default
  pdf_document:
    includes:
      in_header: header.tex
      keep_tex: yes
      latex_engine: xelatex
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## 1 Preparation
### 1.1 Import

```{r 1.1, echo=TRUE, message=FALSE, warning=FALSE, paged.print=FALSE}
source("functions.R")
```

### 1.2 Inputs

Read input files.

```{r 1.2, echo=TRUE}
load("01.cleanData/All/all_shared_sv_dis.RData")
load("01.cleanData/All/all_abun.RData")
load("01.cleanData/All/all_phen.RData")
load("01.cleanData/All/all_fmetab.RData")
info    <- read.table("01.cleanData/Info/Informative_species_information.tsv",
                      sep = "\t", header = T, stringsAsFactors = F)
load("01.cleanData/All/distMat/all_msv_dist_std.RData")
```

## 2 Microbiome PCoA

### 2.1 SV PCoA
```{r 2.1}
## based on shared sv
all_msv_dist_std_avg<-all_shared_sv_dis

all_msv_dist_std_avg_rmna<-dist_rmna(all_msv_dist_std_avg)
all_phen_sv_rmna <- all_phen[match(rownames(all_msv_dist_std_avg_rmna), all_phen$ID),]

all_sv_avg_dist_mds<-cmdscale(all_msv_dist_std_avg_rmna, k=5, eig = T)
all_sv_avg_dist_pcoa <- data.frame(all_sv_avg_dist_mds$points)
#all_phen_sv_rmna$IBD<-factor(all_phen_sv_rmna$IBD, levels = c(0,1))

p_all_sv_pcoa<-ggplot(all_sv_avg_dist_pcoa,aes(X1,X2))+
  geom_point(aes(color = all_phen_sv_rmna$Cohort),size = 2,alpha = 0.5)+
  stat_ellipse(aes( fill = all_phen_sv_rmna$Cohort, color = all_phen_sv_rmna$Cohort),
               type = "norm",linetype = 2,geom = "polygon",alpha = 0.05,show.legend = F)+
  xlab(paste("PCo1=",round(all_sv_avg_dist_mds$eig[1],digits = 2),"%",sep = ""))+
  ylab(paste("PCo2=",round(all_sv_avg_dist_mds$eig[2],digits = 2),"%",sep = ""))+
  scale_color_manual(name=NULL, 
                     breaks = c("LLD", "IBD"),
                     labels = c("LLD", "IBD"),
                     values = mycolor2_blue_red)+
  scale_fill_manual(name=NULL, 
                     breaks = c("LLD", "IBD"),
                     labels = c("LLD", "IBD"),
                     values = mycolor2_blue_red)+
  theme(plot.subtitle = element_text(vjust = 1), 
        plot.caption = element_text(vjust = 1), 
        axis.line.x =  element_line(),
        axis.line.y = element_line(),
        legend.position = 'bottom',
        legend.title = element_blank(),
        legend.key = element_rect(fill = NA), 
        panel.grid.major = element_line(colour = NA),
        panel.grid.minor = element_line(colour = NA),
        panel.background = element_rect(fill = NA))
p_all_sv_pcoa<-ggExtra::ggMarginal(p_all_sv_pcoa, type = "histogram", groupColour = F, groupFill = TRUE,
                                   xparams = list(bins = 60, alpha = 0.5,position = 'identity', color = 'white'),
                                   yparams = list(bins = 60, alpha = 0.5,position = 'identity', color = 'white'))


wilcox.test(all_sv_avg_dist_pcoa$X1~all_phen_sv_rmna$Cohort)
wilcox.test(all_sv_avg_dist_pcoa$X2~all_phen_sv_rmna$Cohort)

if(!dir.exists("05.Microbiome_population_structure")){dir.create("05.Microbiome_population_structure")}

pdf("05.Microbiome_population_structure/all_sv_PCoA.pdf", width = 5, height = 5)
print(p_all_sv_pcoa)
dev.off()
```


### 2.2 Abundance PCoA

```{r 2.2}
all_abun_dist <- as.matrix(vegdist(na.omit(as.data.frame(all_abun)),method = "canberra"))
all_phen_abun_sv_rmna <- all_phen[match(rownames(all_abun_dist),all_phen$ID),]
#all_phen_abun_sv_rmna$IBD<-factor(all_phen_abun_sv_rmna$IBD, levels = c(0,1))

all_abun_dist_mds<-cmdscale(all_abun_dist, k=5, eig = T)
all_abun_dist_pcoa <- data.frame(all_abun_dist_mds$points)


p_all_abun_pcoa<-ggplot(all_abun_dist_pcoa,aes(X1,X2, color = all_phen_abun_sv_rmna$Cohort))+
  geom_point(size = 2,alpha = 0.5)+
  stat_ellipse(aes(group = all_phen_abun_sv_rmna$Cohort,
                   fill = all_phen_abun_sv_rmna$Cohort,
                   color = all_phen_abun_sv_rmna$Cohort),
               type = "norm",linetype = 2,geom = "polygon",alpha = 0.05,show.legend = F)+
  xlab(paste("PCo1=",round(all_abun_dist_mds$eig[1],digits = 2),"%",sep = ""))+
  ylab(paste("PCo2=",round(all_abun_dist_mds$eig[2],digits = 2),"%",sep = ""))+
  scale_color_manual(name=NULL, 
                     breaks = c("LLD", "IBD"),
                     labels = c("LLD", "IBD"),
                     values = mycolor2_blue_red)+
  scale_fill_manual(name=NULL, 
                    breaks = c("LLD", "IBD"),
                    labels = c("LLD", "IBD"),
                    values = mycolor2_blue_red)+
  theme(plot.subtitle = element_text(vjust = 1), 
        plot.caption = element_text(vjust = 1), 
        axis.line.x =  element_line(),
        axis.line.y = element_line(),
        legend.position = 'bottom',
        legend.title = element_blank(),
        legend.key = element_rect(fill = NA), 
        panel.grid.major = element_line(colour = NA),
        panel.grid.minor = element_line(colour = NA),
        panel.background = element_rect(fill = NA))
p_all_abun_pcoa<-ggExtra::ggMarginal(p_all_abun_pcoa, type = "histogram", groupColour = F, groupFill = TRUE,
                                     xparams = list(bins = 60, alpha = 0.5,position = 'identity', color = 'white'),
                                     yparams = list(bins = 60, alpha = 0.5,position = 'identity', color = 'white'))

pdf("05.Microbiome_population_structure/all_abun_PCoA.pdf", width = 5, height = 5)
print(p_all_abun_pcoa)
dev.off()
```


### 2.3 Effect of abundance on population structure of SV

```{r 2.3, eval=FALSE}
all_sv_abun_inter<-intersect(all_phen_sv_rmna$ID, all_phen_abun_sv_rmna$ID)
all_inter_basic_rmna<-all_phen[match(all_sv_abun_inter, all_phen$ID),]
all_inter_sv_pc   <- all_sv_avg_dist_pcoa[match(all_sv_abun_inter,rownames(all_sv_avg_dist_pcoa)),]
all_inter_abun_pc <- all_sv_avg_dist_pcoa[match(all_sv_abun_inter,rownames(all_abun_dist_pcoa)),]
all_inter_sv_abun_pc<-cbind(all_inter_sv_pc, all_inter_abun_pc)
colnames(all_inter_sv_abun_pc)<-c("SV_PC1", "SV_PC2", "SV_PC3", "SV_PC4", "SV_PC5",
                                  "Abun_PC1", "Abun_PC2", "Abun_PC3", "Abun_PC4", "Abun_PC5")
all_inter_basic_rmna$IBD<-factor(all_inter_basic_rmna$IBD, levels = c(0,1))
## Linear model: Remove confounding effect of abundance
# SV PC1
all_sv_pc1_abun_pc_lm.res<-lm(SV_PC1~Abun_PC1+Abun_PC2+Abun_PC3+Abun_PC4+Abun_PC5, data = all_inter_sv_abun_pc)
Abun_PC1_resi<-all_sv_pc1_abun_pc_lm.res$residuals
wilcox.test(Abun_PC1_resi~all_inter_basic_rmna$IBD[match(names(all_sv_pc1_abun_pc_lm.res$residuals), all_inter_basic_rmna$ID)]) # p-value = 1.496e-06
# SV PC2
all_sv_pc2_abun_pc_lm.res<-lm(SV_PC2~Abun_PC1+Abun_PC2+Abun_PC3+Abun_PC4+Abun_PC5, data = all_inter_sv_abun_pc)
Abun_PC2_resi<-all_sv_pc2_abun_pc_lm.res$residuals
wilcox.test(Abun_PC2_resi~all_inter_basic_rmna$IBD[match(names(all_sv_pc2_abun_pc_lm.res$residuals), all_inter_basic_rmna$ID)]) # p-value = 0.001821

## PERMANOVA
covar<-c("host_Sex", "host_Age", "host_BMI", "Read_count","IBD")
all_covar.rmna<-all_phen[,covar] %>% na.omit
all_sv_dist_abun_inter<-intersect(rownames(all_msv_dist_std_avg_rmna), rownames(all_abun_dist_pcoa)) %>%
  intersect(., rownames(all_covar.rmna))
all_inter_sv_dist<-all_msv_dist_std_avg_rmna[match(all_sv_dist_abun_inter,rownames(all_msv_dist_std_avg_rmna)),
                                             match(all_sv_dist_abun_inter,colnames(all_msv_dist_std_avg_rmna))]
all_inter_abun_pc<-all_abun_dist_pcoa[match(all_sv_dist_abun_inter, rownames(all_abun_dist_pcoa)),]
all_inter_basic<-all_phen[match(all_sv_dist_abun_inter, all_phen$ID),]
all_inter_abun_pc_basic<-cbind(all_inter_abun_pc,all_inter_basic[,covar])

all_sv_dist.adonis<-adonis(as.dist(all_inter_sv_dist)~., data = all_inter_abun_pc_basic)
save(all_sv_dist.adonis, file = "05.Microbiome_population_structure/all_sv_dist.adonis.RData")
```


## 3 Effect of basic phenotypes

### 3.1 PERMANOVA 

```{r 3.1, eval=FALSE}
demo_factors<-c("host_Sex", "host_Age", "host_BMI", "Read_count","X1","X2","X3","X4","X5","IBD")
all_phen_sv_p_adonis.df<-as.data.frame(matrix(NA, nrow = length(demo_factors), ncol = 5))

for (i in 1:length(demo_factors)) {
  #i<-1
  all_phen_sv_p.adonis.i<-adonis(
    as.dist(all_inter_sv_dist)~., #all_msv_dist_std_avg_rmna
    data = as.data.frame(all_inter_abun_pc_basic[,demo_factors[i]])) #as.data.frame(all_phen_sv_rmna[,demo_factors[i]])
  
  all_phen_sv_p_adonis.df[i,c(1:3)]<-c(demo_factors[i],
                                        all_phen_sv_p.adonis.i$aov.tab$R2[1], 
                                        all_phen_sv_p.adonis.i$aov.tab$`Pr(>F)`[1])
}

all_phen_sv_p.adonis.comb<-adonis(as.dist(all_inter_sv_dist)~host_Sex+host_Age+host_BMI+Read_count+X1+X2+X3+X4+X5+IBD,
                                  data = all_inter_abun_pc_basic)
all_phen_sv_p.adonis.comb.mat<-all_phen_sv_p.adonis.comb$aov.tab[c(1:length(demo_factors)),] %>%
  .[match(demo_factors, rownames(.)),]

all_phen_sv_p_adonis.df[,4]<-all_phen_sv_p.adonis.comb.mat$R2
all_phen_sv_p_adonis.df[,5]<-all_phen_sv_p.adonis.comb.mat$`Pr(>F)`
colnames(all_phen_sv_p_adonis.df)<-c("Factor", 
                                      "Individual_R2", "Individual_p",
                                      "Combined_R2", "Combined_p")
all_phen_sv_p_adonis.df$Cumulative_R2<-cumsum(all_phen_sv_p_adonis.df$Combined_R2)
all_phen_sv_p_adonis.df$Factor[5:9]<-c("abun_PC1", "abun_PC2","abun_PC3","abun_PC4","abun_PC5")
write.table(all_phen_sv_p_adonis.df, "05.Microbiome_population_structure/all_phen_sv_p_adonis.tsv",
            sep = "\t", col.names = T, row.names = F, quote = F)
save(all_phen_sv_p_adonis.df, file = "05.Microbiome_population_structure/all_phen_sv_p_adonis.df.RData")
```

### 3.2 Visulization

```{r 3.2, fig.width = 4, fig.height = 3}
demo_factors<-c("host_Sex", "host_Age", "host_BMI", "Read_count","abun_PC1", "abun_PC2","abun_PC3","abun_PC4","abun_PC5","IBD")

## ProPortion
load("05.Microbiome_population_structure/all_phen_sv_p_adonis.df.RData")
knitr::kable(all_phen_sv_p_adonis.df)

all_phen_sv_p_adonis.df.long<-gather(all_phen_sv_p_adonis.df[, c(1,2,6)], 'R2', 'Value', -1)
all_phen_sv_p_adonis.df.long$Value<-as.numeric(all_phen_sv_p_adonis.df.long$Value)
all_phen_sv_p_adonis.df.long$R2 <- factor(all_phen_sv_p_adonis.df.long$R2, 
                                           levels = c("Individual_R2", "Cumulative_R2"))

p_all_phen_sv_p_adonis<-ggplot(all_phen_sv_p_adonis.df.long, aes(Factor, Value,group = R2))+
  geom_bar(aes(fill = R2),position =position_dodge(),stat="identity")+
  xlab(NULL)+
  ylab(bquote("R"^2))+
  scale_x_discrete(limits = demo_factors)+
  scale_y_continuous(expand = c(0,0))+
  scale_fill_manual(name=NULL,
                    breaks = c("Individual_R2", "Cumulative_R2"),
                    labels = c(bquote("Univariate R"^2),bquote("Cumulative R"^2) ),
                    values = mycolor2_green_blue)+
  theme(plot.subtitle = element_text(vjust = 1), 
        plot.caption = element_text(vjust = 1), 
        plot.margin=unit(c(2,1,1,3),"cm"),
        axis.line.x =  element_line(),
        axis.line.y = element_line(),
        axis.text.x = element_text(angle = 45, hjust = 1),
        legend.position="top",
        legend.key = element_rect(fill = NA), 
        legend.text = element_text(size = 10),
        panel.grid.major = element_line(colour = NA,linetype = "dashed",size = 0.25),
        panel.grid.minor = element_line(colour = NA), 
        panel.background = element_rect(fill = NA))

pdf("05.Microbiome_population_structure/basic_sv_adonis_bar.pdf", width = 5, height = 4)
print(p_all_phen_sv_p_adonis)
dev.off()

print(p_all_phen_sv_p_adonis)
```

## 4 Species-level PCoA

```{r 4}
msv_pc_cum0.6<-NA
msv_pc_cum0.6_info<-data.frame(matrix(NA, nrow = nrow(info), ncol = 3))
for (i in c(1:nrow(info))) {
  #  i <- 1
  
  msv_pc_res_i<-get_PCs(all_msv_dist_std[[i]],eig.cutoff = 0.6,pc.cutoff = 10)
  msv_pc_i<-msv_pc_res_i$PCoA
  msv_pc_i<-msv_pc_i[match(rownames(all_phen), rownames(msv_pc_i)),]
  rownames(msv_pc_i) <- rownames(all_phen)
  colnames(msv_pc_i)<-paste(info$organism[i],"_PC", c(1:ncol(msv_pc_i)),sep = "")
  
  msv_pc_cum0.6<-cbind(msv_pc_cum0.6, msv_pc_i)
  msv_pc_cum0.6_info[i,]<-c(info$organism[i], msv_pc_res_i$PC_num, msv_pc_res_i$Total_eig)
}
msv_pc_cum0.6<-msv_pc_cum0.6[,-1]

## all 
write.table(msv_pc_cum0.6, "01.cleanData/All/all_msv_pc_cum0.6.tsv",sep = "\t", quote = F)
save(msv_pc_cum0.6, file = "01.cleanData/All/msv_pc_cum0.6.RData")

## lld
lld_msv_pc_cum0.6<-msv_pc_cum0.6[all_phen$Cohort=="LLD",]
save(lld_msv_pc_cum0.6, file = "01.cleanData/LLD/lld_msv_pc_cum0.6.RData")

## 300OB
ibd_msv_pc_cum0.6<-msv_pc_cum0.6[all_phen$Cohort=="IBD",]
save(ibd_msv_pc_cum0.6, file = "01.cleanData/IBD/ibd_msv_pc_cum0.6.RData")

## PC information
colnames(msv_pc_cum0.6_info)<-c("Species", "PC_number", "Explained_variance_proportion")
write.table(msv_pc_cum0.6_info, "01.cleanData/All//all_msv_pc_cum0.6_info.tsv",col.names = T, row.names = F,sep = "\t", quote = F)
save(msv_pc_cum0.6_info, file = "01.cleanData/All/msv_pc_cum0.6_info.RData")
```
